// vaccineApp/src/infrastructure/interfaces/parentesco.response.interface.ts
// Generated by https://quicktype.io

export interface ParentescoResponse {
    desde:         number;
    limite:        number;
    relationships: Relationship[];
    total:         number;
}

export interface Relationship {
    _id:    ID;
    name:   string;
    status: boolean;
}

export interface ID {
    $oid: string;
}


// Generated by https://quicktype.io

export interface CreateResponse {
    id:         string;
    name:       string;
    status:     boolean;
    statusCode: number;
    resp:       boolean;
    message:    string;
}

// Generated by https://quicktype.io